Impl navigate to tab screen
