+++ app/src/main/java/com/numero/material_gallery/components/list/ComponentListFragment.kt
@@ -36,7 +36,6 @@ import com.numero.material_gallery.components.sheet.BottomSheetActivity
 import com.numero.material_gallery.components.sheet.ModalBottomSheetActivity
 import com.numero.material_gallery.components.slider.SliderActivity
 import com.numero.material_gallery.components.snackbar.SnackbarActivity
-import com.numero.material_gallery.components.tab.TabActivity
 import com.numero.material_gallery.view.ListItemAdapter
 import kotlinx.android.synthetic.main.fragment_component_list.*
 
@@ -181,7 +180,7 @@ class ComponentListFragment : Fragment(R.layout.fragment_component_list) {
                 startActivity(SwitchActivity.createIntent(requireContext()))
             }
             DesignComponent.TAB -> {
-                startActivity(TabActivity.createIntent(requireContext()))
+                findNavController().navigate(R.id.action_ComponentList_to_Tab)
             }
             DesignComponent.TEXT_FIELDS -> {
                 findNavController().navigate(R.id.action_ComponentList_to_TextField)
